+++ xv6/user.h
@@ -23,8 +23,8 @@ int getpid(void);
 char* sbrk(int);
 int sleep(int);
 int uptime(void);
-float hitrate(void);
-float resethitrate(void);
+int hitrate(void);
+int resethitrate(void);
 
 // ulib.c
 int stat(const char*, struct stat*);
